requirements/dev.txt
django-extensions
Sphinx
sphinx_rtd_theme
coverage
tox
